// app/model/applicationstatusInfo-model.ts
 export class ApplicationstatusInfo {
      user_list: string;
        active_user_count: string;
        dead_locks: string;
        buffer_busy_waits: string;
        query_sessions: string;
        fast_recovery_usage: string;
        log_file_completion: string;
        db_Sequential_read_wait: string;
        direct_path_reads: string;
        db_file_size: string;
        rollbacks: string;
        commits: string;
        up_time: string;
        redo_writes: string;
        parallel_writes: string;
    }